Изменены модели surveys и schedules.
